+++ sample-functions/ApiKeyProtected-Secrets/.gitignore
@@ -1,0 +1,2 @@
+ApiKeyProtected
+
