+++ sample-functions/ApiKeyProtected-Secrets/build.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+echo Building functions/api-key-protected:latest
+docker build --no-cache -t functions/api-key-protected:latest .
