+++ sample-functions/ApiKeyProtected/README.md
@@ -1,6 +1,6 @@
 ### Api-Key-Protected sample
 
-See the [secure secret management guide](../guide/secure_secret_management.md) for instructions on how to use this function.
+To use this sample provide an env variable for the container/service in `secret_api_key`.
 
-When calling via the gateway pass the additional header "X-Api-Key", if it matches the `secret_api_key` value then the function will give access, otherwise access denied.
+Then when calling via the gateway pass the additional header "X-Api-Key", if it matches the `secret_api_key` value then the function will give access, otherwise access denied.
